backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
openai>=1.12.0
httpx>=0.26.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
psutil>=5.9.0
pyperclip>=1.8.2
python-dotenv>=1.0.0
# 向量嵌入
sentence-transformers>=2.2.0
numpy>=1.24.0
faiss-cpu>=1.7.4
# 联网搜索
duckduckgo_search>=4.0.0
beautifulsoup4>=4.12.0
# EvoMap 进化网络
aiohttp>=3.9.0
